Address conflicts review comments
